refactor(cli): split explicit opt-in/migration out of the storage router
Codex P2 (round 6): device_credential_storage.go crossed the <~400 LOC
guideline. The owner-decision path (forceDeviceCredentialFileMode, the
keyring→file migration, the direct pending read) moves to
device_credential_migration.go; the router/probe stays put. No behavior
change.
